requirements.txt
streamlit
python-dotenv
google-genai
PyMuPDF
streamlit
python-dotenv
pandas
openpyxl
pillow
plotly
python-docx
pymupdf
google-generativeai
